incomplete: implement specified flux boundaries in model setup workflow   * refactor MF6model.setup_wel to follow same structure as other basic stress packages (use the _setup_basic_stress_package fn, which combines perimeter boundaries with user input and for the Well Package, potential input from a parent model)   * implement general basic stress package support for MODFLOW-NWT models   * move _setup_basic_stress_package to MFsetupMixin to use same workflow for both MODFLOW versions   * TODO: get MODFLOW-NWT well package setup tests to pass; either adapt or remove any old tests related to specified flux boundaries   * add specified flux perimeter boundary to the config file gallery for a working example   * remove old Tmr class from tmr.py and any other kruft   * update docs, including limitation that specified flux models must be co-linear
