+++ mfsetup/mf6_defaults.yml
@@ -220,9 +220,10 @@ wel:
   output_files:
     lookup_file: '{}_wel_lookup.csv' # output file that maps wel package data to site numbers
     dropped_wells_file: '{}_dropped_wells.csv' # output file that records wells that were dropped during model setup
-  minimum_layer_thickness: 2.
-  external_files: True  # option to write stress_period_data to external files
-  external_filename_fmt: "wel_{:03d}.dat"
+  mfsetup_options:
+    minimum_layer_thickness: 2.
+    external_files: True  # option to write stress_period_data to external files
+    external_filename_fmt: "wel_{:03d}.dat"
